Broke down the Housing Page into several components
