+++ employee-portal/src/App.tsx
@@ -3,7 +3,7 @@ import NavBar from './components/NavBar';
 import Home from './pages/Home';
 import LoginPage from './pages/Login';
 import RegisterPage from './pages/Register';
-import Housing from './pages/housing/Housing';
+import HousingPage from './pages/housing_without_backend/HousingContainer';
 
 function App() {
   return (
@@ -15,7 +15,7 @@ function App() {
           <Route>
             <Route path="login" element={<LoginPage />} />
             <Route path="register" element={<RegisterPage />} />
-            <Route path="housing" element={<Housing />} />
+            <Route path="housing" element={<HousingPage />} />
           </Route>
         </Routes>
       </BrowserRouter>
